Add files via upload
Real-world data rarely come clean. This was a project from udacity where I had to identity some problems with the datasets. I used python and its libraries to gather data from various sources, assess its quality and tidiness and then clean it to be used for further analysis. Kindly note that the data is not fully cleaned but it is clean enough to do some analysis as seen in the notebook. 
With the Tweepy API, data was collected from the tweet archive of Twitter user @dog_rate which is also known as WeRateDogs.